Relocate render, asset management logic to `render` feature directory
