+++ src/lib/features/render/assets.ts
